deletes all Http and uses OpenWeatherMap API

+++ app/src/main/java/com/bitcamp2018/myclimate/model/Settings.java
@@ -7,7 +7,7 @@ package com.bitcamp2018.myclimate.model;
 public class Settings {
     public static boolean allowNoti = true;
     public static boolean isCel = false;
-    public static double hours = 0;
+    public static int hours = 0;
     public static String condition = "";
     public static String temp_range = "";
     public static String current_face = "happy";
